# openapi.yaml
openapi: 3.0.0
info:
  title: HiveFi Protocol API
  version: 1.0.0
  description: API specification for the HiveFi decentralized AI orchestrator.
servers:
  - url: http://localhost:3001
    description: Local development server
paths:
  /api/orchestrate:
    post:
      summary: Delegate a prompt to the HiveFi swarm
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                socketId:
                  type: string
      responses:
        '200':
          description: Successful orchestration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  delegate:
                    type: boolean
                  text:
                    type: string
                  niche:
                    type: string
                  result:
                    type: string
        '400':
          description: Missing prompt
        '401':
          description: Unauthorized
        '500':
          description: Internal error
  /api/registry:
    get:
      summary: List all registered specialists
      responses:
        '200':
          description: List of specialists
  /api/registry/register-endpoint:
    post:
      summary: Register off-chain endpoint for a model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - modelId
                - endpointUrl
              properties:
                modelId:
                  type: string
                endpointUrl:
                  type: string
      responses:
        '200':
          description: Successfully registered
        '400':
          description: Missing required fields
        '500':
          description: Internal error
  /api/ratings:
    post:
      summary: Submit a rating for a specialist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - modelId
                - taskId
                - score
                - niche
              properties:
                modelId:
                  type: string
                taskId:
                  type: string
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
                niche:
                  type: string
      responses:
        '200':
          description: Rating submitted
        '400':
          description: Invalid rating parameters
        '500':
          description: Internal error
  /api/ratings/{modelId}:
    get:
      summary: Get rating summary for a model
      parameters:
        - in: path
          name: modelId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rating summary
        '500':
          description: Internal error
  /api/balances:
    get:
      summary: Get USDC balances
      responses:
        '200':
          description: Balances object
        '500':
          description: Internal error
  /api/stake:
    post:
      summary: Get contract ABI and address for staking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - modelId
                - amount
              properties:
                modelId:
                  type: string
                amount:
                  type: string
      responses:
        '200':
          description: Staking details returned
        '400':
          description: Missing required fields
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key